requirements.txt
asgiref                == 3.2.5     
certifi                ==2020.4.5.1
chardet                == 3.0.4     
cloudinary             == 1.20.0    
defusedxml             == 0.6.0     
Django                 == 3.0.4     
django-allauth         == 0.41.0    
django-cors-headers    == 3.2.1     
django-rest-auth       == 0.9.5     
djangorestframework    == 3.11.0    
djangorestframework-jwt == 1.11.0    
idna                   == 2.9       
oauthlib               == 3.1.0
Pillow                 == 7.0.0
pip                    == 10.0.1
psycopg2               == 2.8.4
PyJWT                  == 1.7.1
python-dotenv          == 0.13.0
python3-openid         == 3.1.0
pytz                   == 2019.3
requests               == 2.23.0
requests-oauthlib      == 1.3.0
setuptools             == 39.0.1
six                    == 1.14.0
sqlparse               == 0.3.1
stripe                 == 2.47.0
urllib3                == 1.25.9